requirements.txt
asgiref==3.8.1
Django==5.1.3
django-axes==7.0.0
django-cors-headers==4.6.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
PyJWT==2.10.0
sqlparse==0.5.2
